+++ backlog4s-core/src/main/scala/com/github/chaabaj/backlog4s/datas/IssueType.scala
@@ -5,7 +5,7 @@ case class IssueType(
   projectId: Id[Project],
   name: String,
   color: RGBColor,
-  displayOrder: Long
+  displayOrder: Int
 )
 
 case class UpdateIssueTypeForm(
