+++ backlog4s-graphql/src/main/scala/com/github/chaabaj/backlog4s/graphql/schemas/GroupSchema.scala
@@ -14,7 +14,7 @@ object GroupSchema extends BacklogSchema[Unit, Group] {
         Field(
           "id",
           IntType,
-          resolve = _.value.id.value.toInt
+          resolve = _.value.id.value
         ),
         Field(
           "name",
