+++ backlog4s-graphql/src/main/scala/com/github/chaabaj/backlog4s/graphql/schemas/PullRequestSummarySchema.scala
@@ -13,22 +13,22 @@ object PullRequestSummarySchema extends BacklogSchema[Unit, PullRequestSummary] 
         Field(
           "id",
           IntType,
-          resolve = _.value.id.value.toInt
+          resolve = _.value.id.value
         ),
         Field(
           "projectId",
           IntType,
-          resolve = _.value.projectId.value.toInt
+          resolve = _.value.projectId.value
         ),
         Field(
           "repositoryId",
           IntType,
-          resolve = _.value.repositoryId.value.toInt
+          resolve = _.value.repositoryId.value
         ),
         Field(
           "number",
           IntType,
-          resolve = _.value.number.toInt
+          resolve = _.value.number
         ),
         Field(
           "summary",
@@ -63,7 +63,7 @@ object PullRequestSummarySchema extends BacklogSchema[Unit, PullRequestSummary] 
         Field(
           "issue",
           OptionType(IntType),
-          resolve = _.value.issue.map(_.id.value.toInt)
+          resolve = _.value.issue.map(_.id.value)
         ),
         Field(
           "baseCommit",
