+++ backlog4s-graphql/src/main/scala/com/github/chaabaj/backlog4s/graphql/schemas/WikiSchema.scala
@@ -13,12 +13,12 @@ object WikiSchema extends BacklogSchema[BacklogRepository, Wiki] {
         Field(
           "id",
           IntType,
-          resolve = _.value.id.value.toInt
+          resolve = _.value.id.value
         ),
         Field(
           "projectId",
           IntType,
-          resolve = _.value.projectId.value.toInt
+          resolve = _.value.projectId.value
         ),
         Field(
           "project",
